+++ src/main/java/ru/xgzzs0ocfyf1/bookshelf/dao/DAO.java
@@ -8,9 +8,9 @@ public interface DAO<T>{
 
     List<T> list();
     void create(T t);
-    Optional<T> get(int id);
-    void update(T t, int id);
-    void delete(int id);
+    Optional<T> get(long id);
+    void update(T t, long id);
+    void delete(long id);
 
 
 }
